phoenix_engine/modern_python/requirements.txt
fastapi
uvicorn
psycopg2-binary
kafka-python
